fix: Error window loading record undefined.
